# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: ProxySieve local control-plane API
  version: 0.1.0
  description: >-
    The authenticated local Admin API exposed by ProxySieve. This contract
    describes the implemented surface; planned resources are intentionally omitted.
servers:
  - url: http://127.0.0.1:9090
    description: Default local-only listener
...
paths:
...
  /api/v1/pools:
    get:
      tags: [Pools]
      parameters: [{$ref: '#/components/parameters/Limit'}, {name: after, in: query, schema: {type: string}}]
      responses:
        '200': {description: Revisioned routing-pool inventory, content: {application/json: {schema: {$ref: '#/components/schemas/PoolPage'}}}}
    post:
      tags: [Pools]
      security: [sessionCookie: [], csrfToken: []]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [pool], properties: {pool: {$ref: '#/components/schemas/PoolInput'}}}}}
      responses:
        '201': {description: Pool inventory record created, content: {application/json: {schema: {$ref: '#/components/schemas/PoolResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {description: Pool ID already exists}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: proxysieve_session}
    csrfToken: {type: apiKey, in: header, name: X-CSRF-Token}
...
  parameters:
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 1000, default: 100}}
...
  responses:
    BadRequest: {description: Request was not accepted, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}}}
...
    PoolInput:
      type: object
      description: Pool metadata accepted on create; omit id to let the server generate it.
      required: [name, strategy, endpoint_ids, fallback_pool_ids, required_tags, min_health_score, max_latency_ns, enabled]
      properties:
        id: {type: string}
        name: {type: string, minLength: 1, maxLength: 256}
        strategy: {type: string, enum: [random, round-robin, weighted-random, least-connections, least-traffic, lowest-latency, highest-health, lowest-cost, cost-aware, sticky]}
        endpoint_ids: {type: array, maxItems: 1000, uniqueItems: true, items: {type: string}}
        fallback_pool_ids: {type: array, maxItems: 1000, uniqueItems: true, items: {type: string}}
        required_tags: {type: array, maxItems: 1000, uniqueItems: true, items: {type: string, minLength: 1, maxLength: 128}}
        country: {type: string, maxLength: 128}
        min_health_score: {type: integer, minimum: 0, maximum: 100}
        max_latency_ns: {type: integer, minimum: 0}
        session_policy: {$ref: '#/components/schemas/SessionPolicy'}
        enabled: {type: boolean}
    SessionPolicy:
      type: object
      required: [strategy, ttl_ns, idle_ttl_ns, max_requests, max_bytes]
      properties:
        strategy: {type: string, enum: [none, explicit, client, destination, client_destination]}
        ttl_ns: {type: integer, minimum: 0, maximum: 31536000000000000}
        idle_ttl_ns: {type: integer, minimum: 0, maximum: 31536000000000000}
        max_requests: {type: integer, minimum: 0}
        max_bytes: {type: integer, minimum: 0}
...
    Pool:
      allOf:
        - {$ref: '#/components/schemas/PoolInput'}
        - {type: object, required: [id]}
    PoolRecord:
      type: object
      required: [pool, revision]
      properties:
        pool: {$ref: '#/components/schemas/Pool'}
        revision: {type: integer, minimum: 1}
    PoolPage:
      type: object
      required: [items, next_after]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/PoolRecord'}}
        next_after: {type: string}
    PoolResponse:
      type: object
      required: [pool, runtime_active, activation, runtime_revision]
      properties:
        pool: {$ref: '#/components/schemas/PoolRecord'}
        runtime_active: {type: boolean}
        activation: {type: string, enum: [active, staged]}
        runtime_revision: {type: integer, minimum: 0}
...